# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
  /api/v1/quotations:
    get:
      tags:
        - Quotations
      summary: List Quotations
      description: Get a paginated list of Quotations with optional filtering and sorting
      operationId: listQuotations
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuotationList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Quotations
      summary: Create Quotation
      description: Create a new Quotation
      operationId: createQuotation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateQuotationRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quotation'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/quotation_items:
    get:
      tags:
        - QuotationItems
      summary: List QuotationItems
      description: Get a paginated list of QuotationItems with optional filtering and sorting
      operationId: listQuotationItems
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuotationItemList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - QuotationItems
      summary: Create QuotationItem
      description: Create a new QuotationItem
      operationId: createQuotationItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateQuotationItemRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QuotationItem'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
    Quotation:
      type: object
      required:
        - id
        - quotation_number
        - company_id
        - customer_id
        - status
        - quotation_date
        - currency
        - subtotal
        - tax_rate
        - tax_amount
        - total
        - metadata
      properties:
        id:
          type: string
          format: uuid
        quotation_number:
          type: string
          maxLength: 40
        company_id:
          type: string
          format: uuid
        branch_id:
          nullable: true
          type: string
          format: uuid
        customer_id:
          type: string
          format: uuid
        status:
          type: string
        quotation_date:
          type: string
          format: date
        valid_until:
          nullable: true
          type: string
          format: date
        currency:
          type: string
          maxLength: 3
        subtotal:
          type: string
        tax_rate:
          type: string
        tax_amount:
          type: string
        total:
          type: string
        notes:
          nullable: true
          type: string
          maxLength: 1000
        metadata:
          type: object
...
    CreateQuotationRequest:
      type: object
      required:
        - quotation_number
        - company_id
        - customer_id
        - status
        - quotation_date
        - currency
        - subtotal
        - tax_rate
        - tax_amount
        - total
        - metadata
      properties:
        quotation_number:
          type: string
          maxLength: 40
        company_id:
          type: string
          format: uuid
        branch_id:
          nullable: true
          type: string
          format: uuid
        customer_id:
          type: string
          format: uuid
        status:
          type: string
        quotation_date:
          type: string
          format: date
        valid_until:
          nullable: true
          type: string
          format: date
        currency:
          type: string
          maxLength: 3
        subtotal:
          type: string
        tax_rate:
          type: string
        tax_amount:
          type: string
        total:
          type: string
        notes:
          nullable: true
          type: string
          maxLength: 1000
        metadata:
          type: object
...
    QuotationList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Quotation'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    QuotationItem:
      type: object
      required:
        - id
        - quotation_id
        - company_id
        - item_id
        - quantity
        - unit_price
        - line_discount
        - line_amount
        - metadata
      properties:
        id:
          type: string
          format: uuid
        quotation_id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        item_id:
          type: string
          format: uuid
        description:
          nullable: true
          type: string
          maxLength: 300
        quantity:
          type: string
        unit_price:
          type: string
        line_discount:
          type: string
        line_amount:
          type: string
        metadata:
          type: object
...
    CreateQuotationItemRequest:
      type: object
      required:
        - quotation_id
        - company_id
        - item_id
        - quantity
        - unit_price
        - line_discount
        - line_amount
        - metadata
      properties:
        quotation_id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        item_id:
          type: string
          format: uuid
        description:
          nullable: true
          type: string
          maxLength: 300
        quantity:
          type: string
        unit_price:
          type: string
        line_discount:
          type: string
        line_amount:
          type: string
        metadata:
          type: object
...
    QuotationItemList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/QuotationItem'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []