# src/presentation/http/quotation_handler.rs
//! Quotation REST handlers
//!
//! Generated by metaphor-schema. Do not edit manually.
//!
//! Uses Axum and backbone-core's BackboneCrudHandler for all 12 CRUD endpoints.

use std::collections::HashMap;
use std::sync::Arc;

use axum::Router;
use serde::{Deserialize, Serialize};
use uuid::Uuid;
use chrono::{NaiveDate};
use rust_decimal::Decimal;

// Backbone framework imports
use backbone_core::http::BackboneCrudHandler;

// Auth integration (optional)
#[cfg(feature = "auth")]
use backbone_auth::middleware::AuthContext;
#[cfg(feature = "auth")]
use backbone_auth::AuthMiddleware;

// Domain imports
use crate::domain::entity::*;
use crate::application::service::{QuotationService, ServiceError};

// DTO imports
use crate::presentation::dto::{CreateQuotationDto, UpdateQuotationDto, PatchQuotationDto, QuotationResponseDto};


/// Application error type
#[derive(Debug, thiserror::Error)]
pub enum QuotationError {
    #[error("Not found: {0}")]
    NotFound(String),
    #[error("Validation error: {0}")]
    Validation(String),
    #[error("Database error: {0}")]
    Database(String),
    #[error("Internal error: {0}")]
    Internal(String),
}

impl From<ServiceError> for QuotationError {
    fn from(err: ServiceError) -> Self {
        match err {
            ServiceError::NotFound => Self::NotFound(err.to_string()),
            ServiceError::Validation(ref msg) => Self::Validation(msg.clone()),
            ServiceError::AlreadyExists(ref msg) => Self::Validation(msg.clone()),
            ServiceError::Repository(ref e) => Self::Database(e.to_string()),
            ServiceError::Internal(ref msg) => Self::Internal(msg.clone()),
        }
    }
}

impl axum::response::IntoResponse for QuotationError {
    fn into_response(self) -> axum::response::Response {
        use axum::http::StatusCode;
        use axum::Json;

        let (status, code) = match &self {
            Self::NotFound(_) => (StatusCode::NOT_FOUND, "QUOTATION_NOT_FOUND"),
            Self::Validation(_) => (StatusCode::BAD_REQUEST, "QUOTATION_VALIDATION_ERROR"),
            Self::Database(_) => (StatusCode::INTERNAL_SERVER_ERROR, "QUOTATION_DATABASE_ERROR"),
            Self::Internal(_) => (StatusCode::INTERNAL_SERVER_ERROR, "QUOTATION_INTERNAL_ERROR"),
        };

        let body = serde_json::json!({
            "success": false,
            "error": code,
            "message": self.to_string(),
        });

        (status, Json(body)).into_response()
    }
}

// =============================================================================
// Route Configuration
// =============================================================================

/// Create Axum router with all 16 Backbone endpoints for Quotation.
///
/// # Routes
///
/// | Method | Path | Description |
/// |--------|------|-------------|
/// | GET | /quotations | List with pagination |
/// | POST | /quotations | Create new |
/// | GET | /quotations/:id | Get by ID |
/// | PUT | /quotations/:id | Full update |
/// | PATCH | /quotations/:id | Partial update |
/// | DELETE | /quotations/:id | Soft delete |
/// | POST | /quotations/bulk | Bulk create |
/// | POST | /quotations/upsert | Upsert |
/// | GET | /quotations/trash | List deleted |
/// | POST | /quotations/:id/restore | Restore |
/// | DELETE | /quotations/empty | Empty trash |
/// | GET | /quotations/:id/deleted | Get deleted by ID |
/// | DELETE | /quotations/trash/:id | Permanent delete from trash |
/// | GET | /quotations/count | Count active entities |
/// | GET | /quotations/trash/count | Count deleted entities |
///
/// # Example
///
/// ```text
/// let service = Arc::new(QuotationService::with_repository(repository));
/// let router = create_quotation_routes(service);
/// ```
pub fn create_quotation_routes(service: Arc<QuotationService>) -> Router {
    BackboneCrudHandler::<QuotationService, Quotation, CreateQuotationDto, UpdateQuotationDto, QuotationResponseDto>::routes(
        service,
        "/quotations",
    )
}

/// Create Axum router with only the read (GET) endpoints for Quotation.
///
/// Safe for public, unauthenticated exposure (e.g., reference data).
/// Mutations must be served separately via `create_quotation_write_routes`,
/// typically wrapped in an auth middleware layer.
pub fn create_quotation_read_routes(service: Arc<QuotationService>) -> Router {
    BackboneCrudHandler::<QuotationService, Quotation, CreateQuotationDto, UpdateQuotationDto, QuotationResponseDto>::read_routes(
        service,
        "/quotations",
    )
}

/// Create Axum router with only the write (mutation) endpoints for Quotation.
///
/// These routes must NOT be publicly exposed. Wrap them with an auth
/// middleware before nesting into the application router.
pub fn create_quotation_write_routes(service: Arc<QuotationService>) -> Router {
    BackboneCrudHandler::<QuotationService, Quotation, CreateQuotationDto, UpdateQuotationDto, QuotationResponseDto>::write_routes(
        service,
        "/quotations",
    )
}

/// Create authenticated routes with auth middleware.
///
/// Requires the `auth` feature flag. The `AuthMiddleware` implementation
/// is responsible for extracting and validating tokens, then providing
/// an `AuthContext` via request extensions.
#[cfg(feature = "auth")]
pub fn create_protected_quotation_routes<A: AuthMiddleware + Send + Sync + 'static>(
    service: Arc<QuotationService>,
    auth: Arc<A>,
) -> Router {
    use axum::middleware;
    use axum::response::IntoResponse;

    let auth_layer = auth.clone();
    create_quotation_routes(service)
        .layer(middleware::from_fn(move |mut req: axum::extract::Request, next: axum::middleware::Next| {
            let auth = auth_layer.clone();
            async move {
                let token = req.headers()
                    .get(axum::http::header::AUTHORIZATION)
                    .and_then(|h| h.to_str().ok())
                    .and_then(|raw| raw.strip_prefix("Bearer ").or_else(|| raw.strip_prefix("bearer ")))
                    .unwrap_or("");
                match auth.authenticate(token).await {
                    Ok(ctx) => {
                        req.extensions_mut().insert(ctx);
                        next.run(req).await
                    }
                    Err(_) => {
                        (axum::http::StatusCode::UNAUTHORIZED,
                         axum::Json(serde_json::json!({
                             "success": false,
                             "error": "unauthorized",
                             "message": "Authentication required"
                         }))
                        ).into_response()
                    }
                }
            }
        }))
}