# src/presentation/http/quotation_item_handler.rs
//! QuotationItem REST handlers
//!
//! Generated by metaphor-schema. Do not edit manually.
//!
//! Uses Axum and backbone-core's BackboneCrudHandler for all 12 CRUD endpoints.

use std::collections::HashMap;
use std::sync::Arc;

use axum::Router;
use serde::{Deserialize, Serialize};
use uuid::Uuid;
use rust_decimal::Decimal;

// Backbone framework imports
use backbone_core::http::BackboneCrudHandler;

// Auth integration (optional)
#[cfg(feature = "auth")]
use backbone_auth::middleware::AuthContext;
#[cfg(feature = "auth")]
use backbone_auth::AuthMiddleware;

// Domain imports
use crate::domain::entity::*;
use crate::application::service::{QuotationItemService, ServiceError};

// DTO imports
use crate::presentation::dto::{CreateQuotationItemDto, UpdateQuotationItemDto, PatchQuotationItemDto, QuotationItemResponseDto};


/// Application error type
#[derive(Debug, thiserror::Error)]
pub enum QuotationItemError {
    #[error("Not found: {0}")]
    NotFound(String),
    #[error("Validation error: {0}")]
    Validation(String),
    #[error("Database error: {0}")]
    Database(String),
    #[error("Internal error: {0}")]
    Internal(String),
}

impl From<ServiceError> for QuotationItemError {
    fn from(err: ServiceError) -> Self {
        match err {
            ServiceError::NotFound => Self::NotFound(err.to_string()),
            ServiceError::Validation(ref msg) => Self::Validation(msg.clone()),
            ServiceError::AlreadyExists(ref msg) => Self::Validation(msg.clone()),
            ServiceError::Repository(ref e) => Self::Database(e.to_string()),
            ServiceError::Internal(ref msg) => Self::Internal(msg.clone()),
        }
    }
}

impl axum::response::IntoResponse for QuotationItemError {
    fn into_response(self) -> axum::response::Response {
        use axum::http::StatusCode;
        use axum::Json;

        let (status, code) = match &self {
            Self::NotFound(_) => (StatusCode::NOT_FOUND, "QUOTATIONITEM_NOT_FOUND"),
            Self::Validation(_) => (StatusCode::BAD_REQUEST, "QUOTATIONITEM_VALIDATION_ERROR"),
            Self::Database(_) => (StatusCode::INTERNAL_SERVER_ERROR, "QUOTATIONITEM_DATABASE_ERROR"),
            Self::Internal(_) => (StatusCode::INTERNAL_SERVER_ERROR, "QUOTATIONITEM_INTERNAL_ERROR"),
        };

        let body = serde_json::json!({
            "success": false,
            "error": code,
            "message": self.to_string(),
        });

        (status, Json(body)).into_response()
    }
}

// =============================================================================
// Route Configuration
// =============================================================================

/// Create Axum router with all 16 Backbone endpoints for QuotationItem.
///
/// # Routes
///
/// | Method | Path | Description |
/// |--------|------|-------------|
/// | GET | /quotation_items | List with pagination |
/// | POST | /quotation_items | Create new |
/// | GET | /quotation_items/:id | Get by ID |
/// | PUT | /quotation_items/:id | Full update |
/// | PATCH | /quotation_items/:id | Partial update |
/// | DELETE | /quotation_items/:id | Soft delete |
/// | POST | /quotation_items/bulk | Bulk create |
/// | POST | /quotation_items/upsert | Upsert |
/// | GET | /quotation_items/trash | List deleted |
/// | POST | /quotation_items/:id/restore | Restore |
/// | DELETE | /quotation_items/empty | Empty trash |
/// | GET | /quotation_items/:id/deleted | Get deleted by ID |
/// | DELETE | /quotation_items/trash/:id | Permanent delete from trash |
/// | GET | /quotation_items/count | Count active entities |
/// | GET | /quotation_items/trash/count | Count deleted entities |
///
/// # Example
///
/// ```text
/// let service = Arc::new(QuotationItemService::with_repository(repository));
/// let router = create_quotation_item_routes(service);
/// ```
pub fn create_quotation_item_routes(service: Arc<QuotationItemService>) -> Router {
    BackboneCrudHandler::<QuotationItemService, QuotationItem, CreateQuotationItemDto, UpdateQuotationItemDto, QuotationItemResponseDto>::routes(
        service,
        "/quotation_items",
    )
}

/// Create Axum router with only the read (GET) endpoints for QuotationItem.
///
/// Safe for public, unauthenticated exposure (e.g., reference data).
/// Mutations must be served separately via `create_quotation_item_write_routes`,
/// typically wrapped in an auth middleware layer.
pub fn create_quotation_item_read_routes(service: Arc<QuotationItemService>) -> Router {
    BackboneCrudHandler::<QuotationItemService, QuotationItem, CreateQuotationItemDto, UpdateQuotationItemDto, QuotationItemResponseDto>::read_routes(
        service,
        "/quotation_items",
    )
}

/// Create Axum router with only the write (mutation) endpoints for QuotationItem.
///
/// These routes must NOT be publicly exposed. Wrap them with an auth
/// middleware before nesting into the application router.
pub fn create_quotation_item_write_routes(service: Arc<QuotationItemService>) -> Router {
    BackboneCrudHandler::<QuotationItemService, QuotationItem, CreateQuotationItemDto, UpdateQuotationItemDto, QuotationItemResponseDto>::write_routes(
        service,
        "/quotation_items",
    )
}

/// Create authenticated routes with auth middleware.
///
/// Requires the `auth` feature flag. The `AuthMiddleware` implementation
/// is responsible for extracting and validating tokens, then providing
/// an `AuthContext` via request extensions.
#[cfg(feature = "auth")]
pub fn create_protected_quotation_item_routes<A: AuthMiddleware + Send + Sync + 'static>(
    service: Arc<QuotationItemService>,
    auth: Arc<A>,
) -> Router {
    use axum::middleware;
    use axum::response::IntoResponse;

    let auth_layer = auth.clone();
    create_quotation_item_routes(service)
        .layer(middleware::from_fn(move |mut req: axum::extract::Request, next: axum::middleware::Next| {
            let auth = auth_layer.clone();
            async move {
                let token = req.headers()
                    .get(axum::http::header::AUTHORIZATION)
                    .and_then(|h| h.to_str().ok())
                    .and_then(|raw| raw.strip_prefix("Bearer ").or_else(|| raw.strip_prefix("bearer ")))
                    .unwrap_or("");
                match auth.authenticate(token).await {
                    Ok(ctx) => {
                        req.extensions_mut().insert(ctx);
                        next.run(req).await
                    }
                    Err(_) => {
                        (axum::http::StatusCode::UNAUTHORIZED,
                         axum::Json(serde_json::json!({
                             "success": false,
                             "error": "unauthorized",
                             "message": "Authentication required"
                         }))
                        ).into_response()
                    }
                }
            }
        }))
}